+++ integrations/aws-v3/main.py
@@ -49,6 +49,11 @@ from aws.core.exporters.codebuild import (
     CodeBuildBuildRunExporter,
     PaginatedBuildRunRequest,
 )
+from aws.core.exporters.codedeploy import (
+    CodeDeployApplicationExporter,
+    PaginatedCodeDeployApplicationRequest,
+)
+from aws.core.exporters.codepipeline import PipelineExporter, PaginatedPipelineRequest
 from aws.core.helpers.utils import is_access_denied_exception
 
 from loguru import logger
@@ -288,6 +293,27 @@ async def resync_codebuild_project_build_run(kind: str) -> ASYNC_GENERATOR_RESYN
         CodeBuildBuildRunExporter,
         PaginatedBuildRunRequest,
         regional=True,
+    )
+    async for batch in service:
+        yield batch
+
+
+@ocean.on_resync(ObjectKind.CODEDEPLOY_APPLICATION)
+async def resync_codedeploy_application(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
+    service = ResyncAWSService(
+        kind,
+        CodeDeployApplicationExporter,
+        PaginatedCodeDeployApplicationRequest,
+        regional=True,
+    )
+    async for batch in service:
+        yield batch
+
+
+@ocean.on_resync(ObjectKind.CODEPIPELINE_PIPELINE)
+async def resync_codepipeline_pipeline(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
+    service = ResyncAWSService(
+        kind, PipelineExporter, PaginatedPipelineRequest, regional=True
     )
     async for batch in service:
         yield batch